Replace multilocusMap with lociMap wherever it doesn't disturbe compatibility with allelelmatch 2.5.4.
When amDataset have two members called multilocus and multilocusMap,
we get problems when the former is set to NULL, then the next attempt
to read it will not return NULL, but the value of the latter!!

This is because R supports using shortened names.

Unfortunatelly the code (in amCluster) relies on setting multilocus to NULL
now and then.

This update passes amregtool::artRun().